+++ CHANGES.rst
@@ -12,11 +12,20 @@ Changes
 
 Changes:
 --------
-- No change.
+- Support `Process` deployment using `OGC` schema (i.e.: `Process` metadata can be provided directly under
+  ``processDescription`` instead of being nested under ``processDescription.process``).
+  This aligns the deployment schema with reference `OGC API - Processes: Deploy, Replace, Undeploy` extension
+  (see |ogc-app-pkg|_ schema).
+  The previous schema for deployment with nested ``process`` field remains supported for backward compatibility.
+
+.. |ogc-app-pkg| replace:: OGC Application Package
+.. _ogc-app-pkg: https://github.com/opengeospatial/ogcapi-processes/blob/master/extensions/deploy_replace_undeploy/standard/openapi/schemas/ogcapppkg.yaml
 
 Fixes:
 ------
-- No change.
+- Fix resolution of the ``default`` field specifier under a list of supported ``formats`` during deployment.
+  For various combinations such as when ``default: True`` format is omitted, or when the default is not ordered first,
+  resolved ``default`` specifically for ``outputs`` definitions would be incorrect.
 
 .. _changes_4.10.0: